ref: Remove EventFactory (#1091)
* ref: Remove EventFactory

* fix: 7.2

* Update src/Client.php


* fix: Codefix

* ref: Use pretty version

* ref: Remove dead code

* ref: CR

* ref: Update composer
